Update Gradle dependencies, introduce data binding and improve SDK versions
This commit upgrades Gradle dependencies, introduces data binding in the Android app module, and updates the SDK versions to modern versions. It also modifies the repositories used in the allprojects block to switch from JCenter to Maven Central. Additionally, the various dependencies are updated, upgrading versions or introducing new dependencies.